Introduced new slash Commands. Split Play into subcommands.
