+++ Discord Bot/SlashCommands/Userinfo.cs
@@ -10,8 +10,10 @@ namespace Discord_Bot.SlashCommands
     class Userinfo : ApplicationCommandModule
     {
         [SlashCommand("userinfo", "Info about given user")]
-        public async Task User(InteractionContext ctx, [Option("user", "mention a user")] DiscordUser member)
+        public async Task User(InteractionContext ctx, [Option("user", "mention a user")] DiscordUser member = null)
         {
+            if(member == null) member = ctx.User;
+            
             var embed = new DiscordEmbedBuilder();
             embed.AddField("User", $"{member.Username}#{member.Discriminator.ToString()}", false)
             .AddField("Discriminator", $"{member.Discriminator.ToString()}", false)
